move shared logic
